move tests with extra reqs to integration folder
