+++ test/integration/multimodal/test_multimodal.py
@@ -1,11 +1,9 @@
-import os
-
 from examples.advanced.multi_modal_pipeline import prepare_multi_modal_data
+from fedot.core.data.multi_modal import MultiModalData
 from fedot.core.pipelines.node import PipelineNode
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.repository.tasks import Task, TaskTypesEnum
 from fedot.core.utils import fedot_project_root
-from fedot.core.data.multi_modal import MultiModalData
 
 
 def generate_multi_modal_pipeline(data: MultiModalData):
@@ -33,27 +31,8 @@ def generate_multi_modal_pipeline(data: MultiModalData):
     return pipeline
 
 
-def generate_multi_task_pipeline():
-    ds_regr = PipelineNode('data_source_table/regr')
-    ds_class = PipelineNode('data_source_table/class')
-
-    scaling_node_regr = PipelineNode('scaling', nodes_from=[ds_regr])
-    scaling_node_class = PipelineNode('scaling', nodes_from=[ds_class])
-
-    dt_class_node = PipelineNode('dt', nodes_from=[scaling_node_class])
-
-    scaling_node_class_2 = PipelineNode('scaling', nodes_from=[dt_class_node])
-
-    root_regr = PipelineNode('dtreg', nodes_from=[scaling_node_regr, scaling_node_class_2])
-
-    initial_pipeline = Pipeline(root_regr)
-
-    return initial_pipeline
-
-
 def test_multi_modal_pipeline():
-    files_path = os.path.join('test', 'data', 'multi_modal')
-    path = os.path.join(str(fedot_project_root()), files_path)
+    path = fedot_project_root().joinpath('test', 'data', 'multi_modal')
     task = Task(TaskTypesEnum.classification)
     images_size = (128, 128)
 
@@ -64,16 +43,3 @@ def test_multi_modal_pipeline():
     prediction = pipeline.predict(fit_data)
 
     assert prediction is not None
-
-
-def test_finding_side_root_node_in_multi_modal_pipeline():
-    reg_root_node = 'dtreg'
-    class_root_node = 'dt'
-
-    pipeline = generate_multi_task_pipeline()
-
-    reg_pipeline = pipeline.pipeline_for_side_task(task_type=TaskTypesEnum.regression)
-    class_pipeline = pipeline.pipeline_for_side_task(task_type=TaskTypesEnum.classification)
-
-    assert reg_pipeline.root_node.operation.operation_type == reg_root_node
-    assert class_pipeline.root_node.operation.operation_type == class_root_node
