+++ test/unit/multimodal/test_multimodal.py
@@ -1,4 +1,7 @@
 from fedot.api.main import Fedot
+from fedot.core.pipelines.node import PipelineNode
+from fedot.core.pipelines.pipeline import Pipeline
+from fedot.core.repository.tasks import TaskTypesEnum
 from test.unit.multimodal.data_generators import get_single_task_multimodal_tabular_data
 
 
@@ -28,3 +31,34 @@ def test_multimodal_api():
 
     assert pipeline is not None
     assert (9, 1) == prediction.shape
+
+
+def generate_multi_task_pipeline():
+    ds_regr = PipelineNode('data_source_table/regr')
+    ds_class = PipelineNode('data_source_table/class')
+
+    scaling_node_regr = PipelineNode('scaling', nodes_from=[ds_regr])
+    scaling_node_class = PipelineNode('scaling', nodes_from=[ds_class])
+
+    dt_class_node = PipelineNode('dt', nodes_from=[scaling_node_class])
+
+    scaling_node_class_2 = PipelineNode('scaling', nodes_from=[dt_class_node])
+
+    root_regr = PipelineNode('dtreg', nodes_from=[scaling_node_regr, scaling_node_class_2])
+
+    initial_pipeline = Pipeline(root_regr)
+
+    return initial_pipeline
+
+
+def test_finding_side_root_node_in_multi_modal_pipeline():
+    reg_root_node = 'dtreg'
+    class_root_node = 'dt'
+
+    pipeline = generate_multi_task_pipeline()
+
+    reg_pipeline = pipeline.pipeline_for_side_task(task_type=TaskTypesEnum.regression)
+    class_pipeline = pipeline.pipeline_for_side_task(task_type=TaskTypesEnum.classification)
+
+    assert reg_pipeline.root_node.operation.operation_type == reg_root_node
+    assert class_pipeline.root_node.operation.operation_type == class_root_node
