+++ test/unit/tasks/test_classification.py
@@ -1,13 +1,6 @@
 import os
 
 import numpy as np
-from golem.utilities.requirements_notificator import warn_requirement
-
-try:
-    import tensorflow as tf
-except ModuleNotFoundError:
-    warn_requirement('tensorflow')
-
 from sklearn.datasets import load_iris, make_classification
 from sklearn.metrics import roc_auc_score as roc_auc
 
@@ -15,28 +8,11 @@ from examples.simple.classification.image_classification_problem import run_imag
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.data.supplementary_data import SupplementaryData
-from fedot.core.operations.evaluation.operation_implementations.models.keras import (
-    FedotCNNImplementation,
-    check_input_array,
-    create_deep_cnn,
-    fit_cnn,
-    predict_cnn
-)
 from fedot.core.pipelines.node import PipelineNode
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum
-from test.unit.common_tests import is_predict_ignores_target
 from test.integration.models.test_model import classification_dataset_with_redundant_features
-
-
-def check_predict_cnn_correct(model, dataset_to_validate):
-    return is_predict_ignores_target(
-        predict_func=predict_cnn,
-        predict_args={'trained_model': model},
-        data_arg_name='predict_data',
-        input_data=dataset_to_validate,
-    )
 
 
 def pipeline_simple() -> Pipeline:
@@ -191,49 +167,3 @@ def test_output_mode_full_probs():
     assert np.array_equal(results_probs.predict, results_default.predict)
     assert results.predict.shape == (len(test_data.target), 2)
     assert results_probs.predict.shape == (len(test_data.target), 1)
-
-
-def test_image_classification_quality():
-    roc_auc_on_valid, _, _ = get_image_classification_data()
-    deviation_composite = roc_auc_on_valid - 0.5
-
-    roc_auc_on_valid, _, _ = get_image_classification_data(composite_flag=False)
-    deviation_simple = roc_auc_on_valid - 0.5
-
-    assert abs(deviation_composite) < 0.25
-    assert abs(deviation_simple) < 0.35
-
-
-def test_cnn_custom_class():
-    cnn_class = FedotCNNImplementation()
-
-    assert cnn_class.params is not None
-    assert type(cnn_class) == FedotCNNImplementation
-
-
-def test_cnn_methods():
-    _, dataset_to_train, dataset_to_validate = get_image_classification_data()
-    image_shape = (28, 28, 1)
-    num_classes = 7
-    epochs = 10
-    batch_size = 128
-
-    cnn_model = create_deep_cnn(input_shape=image_shape,
-                                num_classes=num_classes)
-
-    transformed_x_train, transform_flag = check_input_array(x_train=dataset_to_train.features)
-
-    model = fit_cnn(train_data=dataset_to_train,
-                    model=cnn_model,
-                    epochs=epochs,
-                    batch_size=batch_size)
-
-    prediction = predict_cnn(trained_model=model,
-                             predict_data=dataset_to_validate)
-
-    assert type(cnn_model) == tf.keras.Sequential
-    assert transform_flag is True
-    assert cnn_model.input_shape[1:] == image_shape
-    assert cnn_model.output_shape[1] == num_classes
-    assert type(prediction) == np.ndarray
-    assert check_predict_cnn_correct(model, dataset_to_validate)
